fix: remove redundant code duplication getContentSrcDecorated
